KVM: X86: Move hardware enable/disable logic into VAC
